+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add `collapsed` option in `book.toml` to collapse code blocks by default
 - Add `uncollapsed` flag for code blocks to overrule the `collapsed` option
+- Ensure the "Inputs:" text is in a separate paragraph from the playground
 
 ## v0.1.4